Display logs in native using sliding window approach
* This approach avoids having to load existing logs but is complex and
  is based on assumptions that the logs are sequenced and sorted.

+++ application/apps/indexer/gui/application/src/session/ui/logs_table/mod.rs
@@ -61,7 +61,7 @@ impl TableDelegate for LogsDelegate<'_> {
         self.to_fetch = self
             .session_data
             .main_table
-            .check_fetch(&info.visible_rows, self.session_data.logs_count)
+            .check_fetch(&info.visible_rows, self.session_data.logs_count as usize)
             .map(LineRange::from);
     }
